feat: move more utility files to commons
- ForwardingPainter.kt → commons/ui/components/
- ZonedSwipeModifier.kt → commons/ui/components/
- ToastMsg.kt → commons/ui/components/toasts/ (with typealias)
